acertados os comportamentos de todas as funcionalidades para montagem da apresentação do pitch.
entrega do relatório final do projeto aplicado.

+++ back-end/src/features/catequista/catequista.service.ts
@@ -1,13 +1,18 @@
-import {Injectable} from '@nestjs/common';
+import {BadRequestException, Injectable} from '@nestjs/common';
 
 import {Catequista} from "./catequista";
 import {CatequistaRepository} from "./catequista.repository";
+import {FieldChecker} from "../../core/infra/field.checker";
+import {Util} from "../../core/util/util";
+import {TurmaService} from "../turma/turma.service";
 
 @Injectable()
 export class CatequistaService {
     private repo: CatequistaRepository;
 
-    constructor() {
+    constructor(
+        private readonly turmaService: TurmaService
+    ) {
         this.repo = new CatequistaRepository();
     }
 
@@ -24,18 +29,37 @@ export class CatequistaService {
     }
 
     salvarCatequista(catequista: Catequista): void {
-        this.repo.save(catequista);
+        const validated = this.getValidated(catequista);
+        this.repo.save(validated);
     }
 
     atualizarCatequista(catequista: Catequista) {
-        this.repo.update(catequista);
+        const validated = this.getValidated(catequista);
+        this.repo.update(validated);
     }
 
     deleteCatequista(catequistaId: number) {
+        const turmas = this.turmaService.findByIdCatequista(catequistaId);
+        if(turmas && 0 < turmas.length){
+            throw new BadRequestException('O catequista já possui turmas.');
+        }
         this.repo.delete(catequistaId);
     }
 
     public findByEmail(email: string): Catequista {
         return this.repo.findByEmail(email);
+    }
+
+    private getValidated(catequista: Catequista): Catequista {
+        const {id, idEtapa, casado, endereco, telefoneMovel, telefoneFixo, email, nome} = catequista;
+        FieldChecker.begin(catequista)
+            .checkIfUndefinedOrNull(idEtapa, 'Informe a etapa.')
+            .checkIfUndefinedOrNull(casado, 'Informe o estado civil.')
+            .checkIfUndefinedOrNull(endereco, 'Informe o endereço.')
+            .checkIfUndefinedOrNull(email, 'Informe o email.')
+            .checkIfUndefinedOrNull(nome, 'Informe o nome.')
+            .checkIf(() => !Util.isTelefoneOk(telefoneMovel, 9), 'Informe um telefone móvel válido.')
+            .validate();
+        return {id, idEtapa, casado, endereco, telefoneMovel, telefoneFixo, email, nome} as Catequista;
     }
 }
